+++ media-sample/src/androidTest/java/com/google/android/horologist/mediasample/AudioOffloadTest.kt
@@ -28,6 +28,7 @@ import com.google.android.horologist.audio.AudioOutput
 import com.google.android.horologist.media.data.Media3MediaItemMapper
 import com.google.android.horologist.media.model.MediaItem
 import com.google.android.horologist.media3.flows.waitForPlaying
+import com.google.android.horologist.mediasample.playback.BasePlaybackTest
 import com.google.common.truth.Truth.assertThat
 import com.google.common.truth.Truth.assertWithMessage
 import kotlinx.coroutines.Dispatchers
@@ -35,93 +36,98 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.test.runTest
 import kotlinx.coroutines.withContext
 import kotlinx.coroutines.withTimeout
+import org.junit.Assume.assumeFalse
 import org.junit.Test
 import kotlin.time.Duration.Companion.seconds
 
 class AudioOffloadTest : BasePlaybackTest() {
-    @Test
-    fun testAudioOffload() {
-        runTest(dispatchTimeoutMs = 600000) {
-            val browser = browser()
+    override fun checkSupportedConfig() {
+        super.checkSupportedConfig()
 
-            val mediaItem = Media3MediaItemMapper.map(
-                MediaItem(
-                    id = "1",
-                    title = "Intro - The Way Of Waking Up (feat. Alan Watts)",
-                    uri = "https://storage.googleapis.com/uamp/The_Kyoto_Connection_-_Wake_Up/01_-_Intro_-_The_Way_Of_Waking_Up_feat_Alan_Watts.mp3",
-                    artist = "The Kyoto Connection",
-                    artworkUri = "https://storage.googleapis.com/uamp/The_Kyoto_Connection_-_Wake_Up/art.jpg",
-                )
+        assumeFalse("Offload not supported on emulator", appContainer.isEmulator)
+    }
+
+    @Test
+    fun testAudioOffload() = runTest {
+        val browser = browser()
+
+        val mediaItem = Media3MediaItemMapper.map(
+            MediaItem(
+                id = "1",
+                title = "Intro - The Way Of Waking Up (feat. Alan Watts)",
+                uri = "https://storage.googleapis.com/uamp/The_Kyoto_Connection_-_Wake_Up/01_-_Intro_-_The_Way_Of_Waking_Up_feat_Alan_Watts.mp3",
+                artist = "The Kyoto Connection",
+                artworkUri = "https://storage.googleapis.com/uamp/The_Kyoto_Connection_-_Wake_Up/art.jpg",
             )
+        )
+
+        assertThat(audioOffloadManager.foreground.value).isFalse()
+
+        val output = audioOutputRepository.audioOutput.value
+
+        withContext(Dispatchers.Main) {
+            browser.setMediaItem(mediaItem)
+            browser.prepare()
+            browser.play()
+
+            assertThat(audioOffloadManager.offloadSchedulingEnabled.value).isTrue()
+
+            withTimeout(10.seconds) {
+                browser.waitForPlaying()
+            }
+
+            val format: Format = audioOffloadManager.format.value!!
+
+            val offloadExpected = offloadExpected(output, Build.MODEL, format)
+
+            val flags = checkAdbDumpsys()
+
+            if (offloadExpected) {
+                // This should be always true because ExoPlayer does this
+                val audioFormat = format.audioFormat()
+                val offloadedPlaybackSupported =
+                    AudioManager.isOffloadedPlaybackSupported(
+                        audioFormat,
+                        browser.audioAttributes.audioAttributesV21.audioAttributes
+                    )
+                assertThat(offloadedPlaybackSupported).isTrue()
+
+                val formatSupport = audioSink.getFormatSupport(format)
+                assertThat(formatSupport).isEqualTo(AudioSink.SINK_FORMAT_SUPPORTED_DIRECTLY)
+
+                assertWithMessage("sleepingForOffload")
+                    .that(audioOffloadManager.sleepingForOffload.value)
+                    .isTrue()
+
+                assertThat(flags.contains("AUDIO_OUTPUT_FLAG_COMPRESS_OFFLOAD"))
+
+                assertThat(audioOffloadManager.times.value.enabled).isGreaterThan(0)
+            } else {
+                assertWithMessage("sleepingForOffload")
+                    .that(audioOffloadManager.sleepingForOffload.value)
+                    .isFalse()
+
+                assertThat(!flags.contains("AUDIO_OUTPUT_FLAG_COMPRESS_OFFLOAD"))
+                assertThat(flags.contains("AUDIO_OUTPUT_FLAG_PRIMARY"))
+
+                assertThat(audioOffloadManager.times.value.enabled).isEqualTo(0L)
+            }
 
             assertThat(audioOffloadManager.foreground.value).isFalse()
 
-            val output = audioOutputRepository.audioOutput.value
+            if (offloadExpected) {
+                delay(3000)
 
-            withContext(Dispatchers.Main) {
-                browser.setMediaItem(mediaItem)
-                browser.prepare()
-                browser.play()
-
-                assertThat(audioOffloadManager.offloadSchedulingEnabled.value).isTrue()
-
-                withTimeout(10.seconds) {
-                    browser.waitForPlaying()
-                }
-
-                val format: Format = audioOffloadManager.format.value!!
-
-                val offloadExpected = offloadExpected(output, Build.MODEL, format)
-
-                val flags = checkAdbDumpsys()
-
-                if (offloadExpected) {
-                    // This should be always true because ExoPlayer does this
-                    val audioFormat = format.audioFormat()
-                    val offloadedPlaybackSupported =
-                        AudioManager.isOffloadedPlaybackSupported(
-                            audioFormat,
-                            browser.audioAttributes.audioAttributesV21.audioAttributes
-                        )
-                    assertThat(offloadedPlaybackSupported).isTrue()
-
-                    val formatSupport = audioSink.getFormatSupport(format)
-                    assertThat(formatSupport).isEqualTo(AudioSink.SINK_FORMAT_SUPPORTED_DIRECTLY)
-
-                    assertWithMessage("sleepingForOffload")
-                        .that(audioOffloadManager.sleepingForOffload.value)
-                        .isTrue()
-
-                    assertThat(flags.contains("AUDIO_OUTPUT_FLAG_COMPRESS_OFFLOAD"))
-
-                    assertThat(audioOffloadManager.times.value.enabled).isGreaterThan(0)
-                } else {
-                    assertWithMessage("sleepingForOffload")
-                        .that(audioOffloadManager.sleepingForOffload.value)
-                        .isFalse()
-
-                    assertThat(!flags.contains("AUDIO_OUTPUT_FLAG_COMPRESS_OFFLOAD"))
-                    assertThat(flags.contains("AUDIO_OUTPUT_FLAG_PRIMARY"))
-
-                    assertThat(audioOffloadManager.times.value.enabled).isEqualTo(0L)
-                }
-
-                assertThat(audioOffloadManager.foreground.value).isFalse()
-
-                if (offloadExpected) {
-                    delay(3000)
-
-                    val times = audioOffloadManager.times.value
-                    assertThat(times.enabled).isGreaterThan(times.disabled)
-                } else {
-                    // In logs
-                    // AudioTrack init failed 0 Config(8000, 12, 2000000) (recoverable)
-                }
-
-                assertThat(browser.isPlaying).isTrue()
-
-                browser.stop()
+                val times = audioOffloadManager.times.value
+                assertThat(times.enabled).isGreaterThan(times.disabled)
+            } else {
+                // In logs
+                // AudioTrack init failed 0 Config(8000, 12, 2000000) (recoverable)
             }
+
+            assertThat(browser.isPlaying).isTrue()
+
+            browser.stop()
         }
     }
